создали фикстуры авторизации и новый тест
