+++ dm_api_account/apis/account_api.py
@@ -15,6 +15,15 @@ class AccountApi(RestClient):
         return response
 
 
+    def get_v1_account(self, **kwargs):
+        """
+        Get current user
+        :return:
+        """
+        response = self.get(path=f'/v1/account', **kwargs)
+        return response
+
+
     def put_v1_account_token(self, token):
         """
         Activate registered user
