+++ helpers/account_helper.py
@@ -85,6 +85,13 @@ class AccountHelper:
         self.mailhog = mailhog
 
 
+    def auth_client(self, login: str, password: str):
+
+        response = self.dm_account_api.login_api.post_v1_account_login(json_data={'login': login, 'password': password})
+        token = {"x-dm-auth-token": response.headers["x-dm-auth-token"]}
+        self.dm_account_api.account_api.set_headers(token)
+        self.dm_account_api.login_api.set_headers(token)
+
     def register_new_user(self, login: str, password: str, email: str):
 
         json_data = {
