+++ restclient/client.py
@@ -9,11 +9,14 @@ from restclient.configuration import Configuration
 class RestClient:
     def __init__(self, configuration: Configuration):
         self.host = configuration.host
-        self.headers = configuration.headers
+        self.set_headers(configuration.headers)
         self.desable_log = configuration.disable_log
         self.session = session()
         self.log = structlog.get_logger(__name__).bind(service='api')
 
+    def set_headers(self, headers):
+        if headers:
+            self.session.headers.update(headers)
 
     def post(self, path, **kwargs):
         return self._send_request(method='POST', path=path, **kwargs)
