+++ tests/conftest.py
@@ -1,0 +1,59 @@
+from collections import namedtuple
+from datetime import datetime
+from json import loads
+
+import pytest
+
+from helpers.account_helper import AccountHelper, time_it
+from restclient.configuration import Configuration as MailhogConfiguration
+from restclient.configuration import Configuration as DmApiConfiguration
+from services.dm_api_account import DMApiAccount
+from services.api_mailhog import MailHogApi
+import structlog
+
+
+structlog.configure(
+    processors=[
+        structlog.processors.JSONRenderer(indent=4,
+                                          ensure_ascii=True,
+                                          # sort_keys=True
+                                          )
+    ]
+)
+
+@pytest.fixture(scope='session')
+def mailhog_api():
+    mailhog_configuration = MailhogConfiguration(host='http://5.63.153.31:5025')
+    mailhog_client = MailHogApi(configuration=mailhog_configuration)
+    return mailhog_client
+
+@pytest.fixture(scope='session')
+def account_api():
+    dm_api_configuration = DmApiConfiguration(host='http://5.63.153.31:5051', disable_log=False)
+    account = DMApiAccount(configuration=dm_api_configuration)
+    return account
+
+@pytest.fixture(scope='session')
+def account_helper(account_api, mailhog_api):
+    account_helper = AccountHelper(dm_account_api=account_api, mailhog=mailhog_api)
+    return account_helper
+
+@pytest.fixture(scope='session')
+def auth_account_helper(mailhog_api):
+    dm_api_configuration = DmApiConfiguration(host='http://5.63.153.31:5051', disable_log=False)
+    account = DMApiAccount(configuration=dm_api_configuration)
+    account_helper = AccountHelper(dm_account_api=account, mailhog=mailhog_api)
+    account_helper.auth_client(login="Good_15_03_2025_22_21%S", password='123456789')
+    return account_helper
+
+@pytest.fixture
+def prepare_user():
+    now = datetime.now()
+    data = now.strftime("%d_%m_%Y_%H_%M%_%S")
+    login = f'Good_{data}'
+    password = '123456789'
+    email = f'{login}@mail.ru'
+    User = namedtuple("User", ["login", "password", "email"])
+    user = User(login=login, password=password, email=email)
+    return user
+
